refactor(canvas): extract interaction logic to useCanvasInteraction hook
* Extracted complex pointer event handling (`handleDown`, `handleMove`, `handleUp`, `handleDeleteShape`) and state (`dragRef`, `cursorMode`) from `DesignCanvas.jsx` into a new custom hook `useCanvasInteraction.js`.
* Separated view rendering concerns from complex state-machine logic for better maintainability and contributor readability.
* Updated `docs/refactoring/ANALYSIS_AND_PLAN.md` with detailed workflows, Zustand/zundo state mapping, and Go backend interactions based on user request.
